+++ src/test/java/org/sonar/plugins/delphi/coverage/delphicodecoveragetool/DelphiCoverageToolParserTest.java
@@ -46,6 +46,7 @@ class DelphiCoverageToolParserTest {
   private static final String NO_LINE_HITS = BASE_REPORT_PATH + "NoLineHits.xml";
   private static final String INVALID_STRUCTURE = BASE_REPORT_PATH + "InvalidStructure.xml";
   private static final String NORMAL_COVERAGE = BASE_REPORT_PATH + "NormalCoverage.xml";
+  private static final String NORMAL_COVERAGE_PART_2 = BASE_REPORT_PATH + "NormalCoverage2.xml";
 
   private static final String GLOBALS_FILENAME = "Globals.pas";
   private static final String GLOBALS_FILE_KEY = ":" + GLOBALS_FILENAME;
@@ -94,6 +95,16 @@ class DelphiCoverageToolParserTest {
     assertThat(context.lineHits(MAIN_WINDOW_FILE_KEY, 38)).isEqualTo((Integer) 1);
     assertThat(context.lineHits(MAIN_WINDOW_FILE_KEY, 39)).isEqualTo((Integer) 1);
     assertThat(context.lineHits(MAIN_WINDOW_FILE_KEY, 40)).isEqualTo((Integer) 1);
+  }
+
+  @Test
+  void testLineHitsFromDifferentReportsAreMerged() {
+    parser.parse(context, DelphiUtils.getResource(NORMAL_COVERAGE));
+    parser.parse(context, DelphiUtils.getResource(NORMAL_COVERAGE_PART_2));
+
+    assertThat(context.lineHits(GLOBALS_FILE_KEY, 16)).isEqualTo((Integer) 1);
+    assertThat(context.lineHits(GLOBALS_FILE_KEY, 17)).isEqualTo((Integer) 1);
+    assertThat(context.lineHits(GLOBALS_FILE_KEY, 23)).isEqualTo((Integer) 1);
   }
 
   void testReportFileIsIgnored(File file) {